Refactoring: Move the host bridge client out of the protobus-service file. (#6062)
