+++ src/hosts/external/host-bridge-client-manager.ts
@@ -13,7 +13,7 @@ import {
 	WorkspaceServiceClientImpl,
 } from "@generated/hosts/standalone/host-bridge-clients"
 import { HostBridgeClientProvider } from "@/hosts/host-provider-types"
-import { HOSTBRIDGE_PORT } from "@/standalone/protobus-service"
+import { HOSTBRIDGE_PORT } from "@/standalone/hostbridge-client"
 
 /**
  * Manager to hold the gRPC clients for the host bridge. The clients should be re-used to avoid
