+++ src/standalone/cline-core.ts
@@ -8,7 +8,8 @@ import { WebviewProvider } from "@/core/webview"
 import { AuthHandler } from "@/hosts/external/AuthHandler"
 import { HostProvider } from "@/hosts/host-provider"
 import { DiffViewProvider } from "@/integrations/editor/DiffViewProvider"
-import { startProtobusService, waitForHostBridgeReady } from "./protobus-service"
+import { waitForHostBridgeReady } from "./hostbridge-client"
+import { startProtobusService } from "./protobus-service"
 import { log } from "./utils"
 import { extensionContext } from "./vscode-context"
